--- a/sdk/go.mod
+++ b/sdk/go.mod
@@ -5,7 +5,6 @@ go 1.14
 replace go.opentelemetry.io/otel => ../
 
 require (
-	github.com/DataDog/sketches-go v0.0.1
 	github.com/benbjohnson/clock v1.0.3
 	github.com/google/go-cmp v0.5.4
 	github.com/google/gofuzz v1.1.0 // indirect