--- a/sdk/go.mod
+++ b/sdk/go.mod
@@ -7,7 +7,6 @@ replace go.opentelemetry.io/otel => ../
 require (
 	github.com/benbjohnson/clock v1.0.3
 	github.com/google/go-cmp v0.5.4
-	github.com/google/gofuzz v1.1.0 // indirect
 	github.com/stretchr/testify v1.6.1
 	go.opentelemetry.io/otel v0.15.0
 )