--- a/sdk/metric/aggregator/minmaxsumcount/mmsc.go
+++ b/sdk/metric/aggregator/minmaxsumcount/mmsc.go
@@ -46,8 +46,7 @@ var _ export.Aggregator = &Aggregator{}
 var _ aggregation.MinMaxSumCount = &Aggregator{}
 
 // New returns a new aggregator for computing the min, max, sum, and
-// count.  It does not compute quantile information other than Min and
-// Max.
+// count.
 //
 // This type uses a mutex for Update() and SynchronizedMove() concurrency.
 func New(cnt int, desc *metric.Descriptor) []Aggregator {