This is a cleanup of an area in the code that is not well exercised.

- Renames `aggregator/array` to `aggregator/exact`
- Add missing exact timestamps to the `exact` aggregator
- Remove Quantile and Distribution export aggregation interfaces
- Remove DDSketch aggregator; we expect this will return after [more histogram types](https://github.com/open-telemetry/opentelemetry-proto/pull/226/files) are available.

Before:

```
goos: darwin
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/metric
BenchmarkInt64ArrayAdd-8           	 3018822	       369 ns/op	     301 B/op	       2 allocs/op
BenchmarkInt64ArrayHandleAdd-8     	27656120	        55.6 ns/op	      46 B/op	       0 allocs/op
BenchmarkFloat64ArrayAdd-8         	 3263607	       423 ns/op	     298 B/op	       2 allocs/op
BenchmarkFloat64ArrayHandleAdd-8   	27069272	        56.4 ns/op	      47 B/op	       0 allocs/op
```

After:

```
BenchmarkInt64ExactAdd-8           	 1657836	       711 ns/op	     419 B/op	       2 allocs/op
BenchmarkInt64ExactHandleAdd-8     	 6865136	       208 ns/op	     188 B/op	       0 allocs/op
BenchmarkFloat64ExactAdd-8         	 2269024	       566 ns/op	     442 B/op	       2 allocs/op
BenchmarkFloat64ExactHandleAdd-8   	 5736513	       178 ns/op	     180 B/op	       0 allocs/op
```

Although this updates the `exact` aggregator, that code is still not used within this repository. A future PR may add support to the OTLP exporter to export raw point values and timestamps.

Closes #1417