@jmacd 2020-12-21
@open-telemetry/specs-metrics-approvers As I am trying to bring the OTel-Go metrics SDK into line with "Everything we will specify", this is something I'd like to remove. (Note: I'm hoping to relieve the @open-telemetry/go-approvers of this responsibility, it's just code deletion.)