@jmacd 2020-12-21
@open-telemetry/specs-metrics-approvers As I am trying to bring the OTel-Go metrics SDK into line with "Everything we will specify", this Quantile aggregation is something I'd like to remove. I was on the fence about adding timestamps and renaming the Array->Exact aggregator, but I think we want to keep an Exact aggregator that captures timestamps, so this is a small change in the forward direction.

(Note: I'm hoping to relieve the @open-telemetry/go-approvers of this responsibility.)