@krnowak 2020-12-21
The PR looks good, just found some nitpicks.