@bogdandrutu 2021-01-08
@jmacd I am not the maintainer, so my comment is just a suggestion. If I was the maintainer I would have pushed for this :)