@MrAlias 2021-01-08
We talked about splitting this up in the SIG meeting yesterday and came to the decision that given the existing approvals we should progress as is. Though future PRs should follow the guidance you laid out @bogdandrutu.

@jmacd I think this can merge once the [this](https://github.com/open-telemetry/opentelemetry-go/pull/1412#discussion_r552899662) is addressed.