@rmrf 2021-07-05
> Where I can find the reason to remove quantile?

I found out the reason: https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/datamodel.md#summary-legacy